Move elevator to single class and begin sim implementation for elevator
